Move extensions into separate file
