+++ Sources/ReactiveTimelane/Signal+Lane.swift
@@ -25,11 +25,11 @@ public extension Signal {
         let fileName = file.description.components(separatedBy: "/").last!
         let source = "\(fileName):\(line) - \(function)"
         let subscription = Timelane.Subscription(name: name, logger: logger)
-
+        
         if filter.contains(.subscription) {
             subscription.begin(source: source)
         }
-
+        
         return on(
             failed: { error in
                 if filter.contains(.subscription) {
@@ -60,51 +60,5 @@ public extension Signal {
                 subscription.event(value: .value(transform(value)), source: source)
             }
         })
-    }
-}
-
-@available(macOS 10.14, iOS 12, tvOS 12, watchOS 5, *)
-public extension SignalProducer {
-    /// The `lane` operator logs the subscription and its events to the Timelane Instrument.
-    ///
-    ///  - Note: You can download the Timelane Instrument from http://timelane.tools
-    /// - Parameters:
-    ///   - name: A name for the lane when visualized in Instruments
-    ///   - filter: Which events to log subscriptions or data events.
-    ///             For example for a subscription on a subject you might be interested only in data events.
-    ///   - transformValue: An optional closure to format the subscription values for displaying in Instruments.
-    ///                     You can not only prettify the values but also change them completely, e.g. for arrays you can
-    ///                     it might be more useful to report the count of elements if there are a lot of them.
-    ///   - value: The value emitted by the subscription
-    func lane(_ name: String,
-              filter: Timelane.LaneTypeOptions = .all,
-              file: StaticString = #file,
-              function: StaticString = #function,
-              line: UInt = #line,
-              transformValue transform: @escaping (_ value: Value) -> String = { String(describing: $0) },
-              logger: @escaping Timelane.Logger = Timelane.defaultLogger) -> SignalProducer<Value, Error> {
-        lift { $0.lane(name, filter: filter, transformValue: transform, logger: logger) }
-    }
-}
-
-@available(macOS 10.14, iOS 12, tvOS 12, watchOS 5, *)
-public extension Lifetime {
-    /// The `lane` operator logs the lifetime to the Timelane Instrument.
-    ///
-    ///  - Note: You can download the Timelane Instrument from http://timelane.tools
-    /// - Parameters:
-    ///   - name: A name for the lane when visualized in Instruments
-    func lane(_ name: String,
-              file: StaticString = #file,
-              function: StaticString = #function,
-              line: UInt = #line,
-              logger: @escaping Timelane.Logger = Timelane.defaultLogger) {
-        let fileName = file.description.components(separatedBy: "/").last!
-        let source = "\(fileName):\(line) - \(function)"
-        let subscription = Timelane.Subscription(name: name, logger: logger)
-        subscription.begin(source: source)
-        self += observeEnded {
-            subscription.end(state: .completed)
-        }
     }
 }
